+++ app/kitchen-sink-ios.js
@@ -426,7 +426,7 @@ module.exports = function(vueApp) {
       });
     });
   });
-  $$('.login-screen').find('.list-button').on('click', function() {
+  $$(document).on('click', '.login-screen .list-button', function() {
     var username = $$('.login-screen').find('input[name="username"]').val();
     var password = $$('.login-screen').find('input[name="password"]').val();
     myApp.alert('Username: ' + username + ', password: ' + password, function() {