+++ app/kitchen-sink-material.js
@@ -431,7 +431,7 @@ module.exports = function(vueApp) {
       });
     });
   });
-  $$('.login-screen').find('.button').on('click', function() {
+  $$(document).on('click', '.login-screen .button', function() {
     var username = $$('.login-screen').find('input[name="username"]').val();
     var password = $$('.login-screen').find('input[name="password"]').val();
     myApp.alert('Username: ' + username + ', password: ' + password, function() {